Decompose `ddp-server` components (mainly classes) into separate files.
