+++ packages/ddp-server/crossbar_tests.js
@@ -1,3 +1,5 @@
+import Crossbar from "./crossbar.js";
+
 // White box tests of invalidation crossbar matching function.
 // Note: the current crossbar match function is designed specifically
 // to ensure that a modification that targets a specific ID does not
@@ -6,7 +8,7 @@
 // deep meaning to the matching function, and it could be changed later
 // as long as it preserves that property.
 Tinytest.add('livedata - crossbar', function (test) {
-  var crossbar = new DDPServer._Crossbar;
+  var crossbar = new Crossbar();
   test.isTrue(crossbar._matches({collection: "C"},
                                 {collection: "C"}));
   test.isTrue(crossbar._matches({collection: "C", id: "X"},
