+++ packages/ddp-server/ddp-server.js
@@ -1,0 +1,1 @@
+export { DDPServer } from "./namespace.js";
