+++ packages/ddp-server/server_convenience.js
@@ -1,9 +1,12 @@
+import { DDPServer } from "./namespace.js";
+import Server from "./server.js";
+
 if (process.env.DDP_DEFAULT_CONNECTION_URL) {
   __meteor_runtime_config__.DDP_DEFAULT_CONNECTION_URL =
     process.env.DDP_DEFAULT_CONNECTION_URL;
 }
 
-Meteor.server = new Server;
+Meteor.server = new Server();
 
 Meteor.refresh = function (notification) {
   DDPServer._InvalidationCrossbar.fire(notification);
