+++ packages/ddp-server/writefence.js
@@ -1,33 +1,28 @@
-var Future = Npm.require('fibers/future');
+import { DDPServer } from "./namespace.js";
+const Future = Npm.require('fibers/future');
 
 // A write fence collects a group of writes, and provides a callback
 // when all of the writes are fully committed and propagated (all
 // observers have been notified of the write and acknowledged it.)
 //
-DDPServer._WriteFence = function () {
-  var self = this;
+export default class WriteFence {
+  constructor() {
+    var self = this;
 
-  self.armed = false;
-  self.fired = false;
-  self.retired = false;
-  self.outstanding_writes = 0;
-  self.before_fire_callbacks = [];
-  self.completion_callbacks = [];
-};
+    self.armed = false;
+    self.fired = false;
+    self.retired = false;
+    self.outstanding_writes = 0;
+    self.before_fire_callbacks = [];
+    self.completion_callbacks = [];
+  }
 
-// The current write fence. When there is a current write fence, code
-// that writes to databases should register their writes with it using
-// beginWrite().
-//
-DDPServer._CurrentWriteFence = new Meteor.EnvironmentVariable;
-
-_.extend(DDPServer._WriteFence.prototype, {
   // Start tracking a write, and return an object to represent it. The
   // object has a single method, committed(). This method should be
   // called when the write is fully committed and propagated. You can
   // continue to add writes to the WriteFence up until it is triggered
   // (calls its callbacks because all writes have committed.)
-  beginWrite: function () {
+  beginWrite() {
     var self = this;
 
     if (self.retired)
@@ -47,40 +42,40 @@ _.extend(DDPServer._WriteFence.prototype, {
         self._maybeFire();
       }
     };
-  },
+  }
 
   // Arm the fence. Once the fence is armed, and there are no more
   // uncommitted writes, it will activate.
-  arm: function () {
+  arm() {
     var self = this;
     if (self === DDPServer._CurrentWriteFence.get())
       throw Error("Can't arm the current fence");
     self.armed = true;
     self._maybeFire();
-  },
+  }
 
   // Register a function to be called once before firing the fence.
   // Callback function can add new writes to the fence, in which case
   // it won't fire until those writes are done as well.
-  onBeforeFire: function (func) {
+  onBeforeFire(func) {
     var self = this;
     if (self.fired)
       throw new Error("fence has already activated -- too late to " +
                       "add a callback");
     self.before_fire_callbacks.push(func);
-  },
+  }
 
   // Register a function to be called when the fence fires.
-  onAllCommitted: function (func) {
+  onAllCommitted(func) {
     var self = this;
     if (self.fired)
       throw new Error("fence has already activated -- too late to " +
                       "add a callback");
     self.completion_callbacks.push(func);
-  },
+  }
 
   // Convenience function. Arms the fence, then blocks until it fires.
-  armAndWait: function () {
+  armAndWait() {
     var self = this;
     var future = new Future;
     self.onAllCommitted(function () {
@@ -88,9 +83,9 @@ _.extend(DDPServer._WriteFence.prototype, {
     });
     self.arm();
     future.wait();
-  },
+  }
 
-  _maybeFire: function () {
+  _maybeFire() {
     var self = this;
     if (self.fired)
       throw new Error("write fence already activated?");
@@ -118,14 +113,14 @@ _.extend(DDPServer._WriteFence.prototype, {
         _.each(callbacks, invokeCallback);
       }
     }
-  },
+  }
 
   // Deactivate this fence so that adding more writes has no effect.
   // The fence must have already fired.
-  retire: function () {
+  retire() {
     var self = this;
     if (! self.fired)
       throw new Error("Can't retire a fence that hasn't fired.");
     self.retired = true;
   }
-});
+}
